% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catter.R
\name{t.test.catter_a}
\alias{t.test.catter_a}
\title{A method}
\usage{
\method{t}{test.catter_a}(a, b, c)
}
\description{
Note that if this method were exported with a regular export tag, roxygen
would misdiagnose it as method t on class test.catter_a (because t exists as
a generic).
}
